pyproject.toml
[tool.poetry]
name = "limit-order-book"
version = "0.1.0"
description = ""
authors = ["Yahor Sivenkou <egor.sivenkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
setuptools = "^69.0.2"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
celery = "^5.3.6"
django-extensions = "^3.2.3"
ipython = "^8.18.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
